Added file provider with CRUD capabilities.
Project Freeze!

Project Freeze! (refactored)

Refactored ParserImpl and task decomposition started

Refactored and freezing of the project due to the war and, as a result, due to the inability to parse data from government sites.

Refactore

Project Freeze! (Refactore class ParserImpl, MyUtilities, RequestHandler)

Added files which are needed for this project