ParserImpl was described
ParserImpl was refactored

ParserImpl was refactored

RequestHandler refactored

Was Refactored RequestHandler, ParserImpl. An error was found in the algorithm of the method parseMinfinCurrencyCodes in ParserImpl

Was Refactored ParserImpl. Fixed errors in the algorithm of the method parseMinfinCurrencyCodes in ParserImpl

Was Refactored ParserImpl. Fixed errors in the algorithm of the method parseMinfinCurrencyCodes in ParserImpl